style: 优化代码风格，移动代码位置
