+++ laokou-service/laokou-network/laokou-network-domain/src/main/java/org/laokou/network/model/MqttMessage.java
